requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
neo4j==6.1.0
python-dotenv==1.2.2
pytz==2026.1.post1
Werkzeug==3.1.7
